refactor: Moved out mocking if environments yaml file
