+++ tests/test_approval_modules.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 
 from collections.abc import Iterator
 from pathlib import Path
-from typing import Any, Literal
 
 import pytest
 import typer
@@ -16,9 +15,10 @@ from pytest import MonkeyPatch
 
 from cognite_toolkit import _version
 from cognite_toolkit.cdf import build, clean, deploy, main_init
-from cognite_toolkit.cdf_tk.templates import COGNITE_MODULES, iterate_modules, read_yaml_file
+from cognite_toolkit.cdf_tk.templates import COGNITE_MODULES, iterate_modules
 from cognite_toolkit.cdf_tk.utils import CDFToolConfig
 from tests.approval_client import ApprovalCogniteClient
+from tests.utils import mock_read_yaml_file
 
 REPO_ROOT = Path(__file__).parent.parent
 
@@ -33,19 +33,16 @@ def find_all_modules() -> Iterator[Path]:
         yield pytest.param(module, id=f"{module.parent.name}/{module.name}")
 
 
-def mock_read_yaml_file(module_path: Path, monkeypatch: MonkeyPatch) -> None:
-    def fake_read_yaml_file(
-        filepath: Path, expected_output: Literal["list", "dict"] = "dict"
-    ) -> dict[str, Any] | list[dict[str, Any]]:
-        if filepath.name == "environments.yaml":
-            return {
+def mock_environments_yaml_file(module_path: Path, monkeypatch: MonkeyPatch) -> None:
+    return mock_read_yaml_file(
+        {
+            "environments.yaml": {
                 "dev": {"project": "pytest-project", "type": "dev", "deploy": [module_path.name]},
                 "__system": {"cdf_toolkit_version": _version.__version__},
             }
-        return read_yaml_file(filepath, expected_output)
-
-    monkeypatch.setattr("cognite_toolkit.cdf_tk.templates.read_yaml_file", fake_read_yaml_file)
-    monkeypatch.setattr("cognite_toolkit.cdf.read_yaml_file", fake_read_yaml_file)
+        },
+        monkeypatch,
+    )
 
 
 @pytest.mark.parametrize("module_path", list(find_all_modules()))
@@ -59,7 +56,7 @@ def test_deploy_module_approval(
     init_project: Path,
     data_regression,
 ) -> None:
-    mock_read_yaml_file(module_path, monkeypatch)
+    mock_environments_yaml_file(module_path, monkeypatch)
 
     build(
         typer_context,
@@ -104,7 +101,7 @@ def test_deploy_dry_run_module_approval(
     typer_context: typer.Context,
     init_project: Path,
 ) -> None:
-    mock_read_yaml_file(module_path, monkeypatch)
+    mock_environments_yaml_file(module_path, monkeypatch)
 
     build(
         typer_context,
@@ -147,7 +144,7 @@ def test_clean_module_approval(
     typer_context: typer.Context,
     data_regression,
 ) -> None:
-    mock_read_yaml_file(module_path, monkeypatch)
+    mock_environments_yaml_file(module_path, monkeypatch)
 
     main_init(
         typer_context,
